chore: refactor action value
